Fix type conversion warning.

+++ zwm.c
@@ -196,8 +196,8 @@ static void motion_notify(xcb_generic_event_t *event)
 	xcb_query_pointer_reply_t *pointer;
 	xcb_get_geometry_cookie_t geom_cookie;
 	xcb_get_geometry_reply_t *geom;
-	uint32_t x;
-	uint32_t y;
+	int16_t x;
+	int16_t y;
 	uint32_t vals[2];
 
 	UNUSED(event);
